refactor json object operators
